+++ src/d-facility.f
@@ -1,6 +1,5 @@
 
 \ The following words are implemented internally:
-\
 \ AT-XY KEY? PAGE
 
 .( ... end of d-facility.f ) cr