+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 v0.3 TBD
 - DOUBLE: add `D0<`, `D0=`, `D<`, `D=`.
+- FACILITY: add `+FIELD`, `BEGIN-STRUCTURE`, `END-STRUCTURE`.
 
 v0.2 alpha 2023-03-20
 - Add `COMPILE-ONLY`.